Speed up AINode CI by consolidating tests and caching PyInstaller output
- Merge 5 AINode IT test classes (DeviceManage, ModelManage, CallInference,
  Forecast, InstanceManagement) into AINodeSharedClusterIT that shares a
  single 1C1D1A cluster, reducing cluster startups from 8 to 3 (~20min saved)
- Convert AINodeClusterConfigIT from @Before/@After to @BeforeClass/@AfterClass,
  merging both dialect tests into one method to eliminate a redundant cluster restart
- Add hash-based dist caching to build_binary.py: computes SHA256 of AINode source
  files and skips PyInstaller rebuild when source hasn't changed (~11min saved)
- Cache stored at ~/.cache/iotdb-ainode-build/dist-cache/, survives mvn clean
