--- integration-test/src/test/java/org/apache/iotdb/ainode/it/AINodeDeviceManageIT.java
@@ -1,96 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.apache.iotdb.ainode.it;
-
-import org.apache.iotdb.it.env.EnvFactory;
-import org.apache.iotdb.it.framework.IoTDBTestRunner;
-import org.apache.iotdb.itbase.category.AIClusterIT;
-import org.apache.iotdb.itbase.env.BaseEnv;
-
-import org.junit.AfterClass;
-import org.junit.Assert;
-import org.junit.BeforeClass;
-import org.junit.Test;
-import org.junit.experimental.categories.Category;
-import org.junit.runner.RunWith;
-
-import java.sql.Connection;
-import java.sql.ResultSet;
-import java.sql.ResultSetMetaData;
-import java.sql.SQLException;
-import java.sql.Statement;
-import java.util.Arrays;
-import java.util.LinkedList;
-import java.util.List;
-
-import static org.apache.iotdb.ainode.utils.AINodeTestUtils.checkHeader;
-import static org.apache.iotdb.ainode.utils.AINodeTestUtils.prepareDataInTable;
-import static org.apache.iotdb.ainode.utils.AINodeTestUtils.prepareDataInTree;
-
-@RunWith(IoTDBTestRunner.class)
-@Category({AIClusterIT.class})
-public class AINodeDeviceManageIT {
-
-  @BeforeClass
-  public static void setUp() throws Exception {
-    // Init 1C1D1A cluster environment
-    EnvFactory.getEnv().initClusterEnvironment(1, 1);
-    prepareDataInTree();
-    prepareDataInTable();
-  }
-
-  @AfterClass
-  public static void tearDown() throws Exception {
-    EnvFactory.getEnv().cleanClusterEnvironment();
-  }
-
-  @Test
-  public void showAIDeviceTestInTree() throws SQLException {
-    try (Connection connection = EnvFactory.getEnv().getConnection(BaseEnv.TREE_SQL_DIALECT);
-        Statement statement = connection.createStatement()) {
-      showAIDevicesTest(statement);
-    }
-  }
-
-  @Test
-  public void showAIDeviceTestInTable() throws SQLException {
-    try (Connection connection = EnvFactory.getEnv().getConnection(BaseEnv.TABLE_SQL_DIALECT);
-        Statement statement = connection.createStatement()) {
-      showAIDevicesTest(statement);
-    }
-  }
-
-  private void showAIDevicesTest(Statement statement) throws SQLException {
-    final String showSql = "SHOW AI_DEVICES";
-    final List<String> expectedDeviceIdList = new LinkedList<>(Arrays.asList("0", "1", "cpu"));
-    final List<String> expectedDeviceTypeList =
-        new LinkedList<>(Arrays.asList("cuda", "cuda", "cpu"));
-    try (ResultSet resultSet = statement.executeQuery(showSql)) {
-      ResultSetMetaData resultSetMetaData = resultSet.getMetaData();
-      checkHeader(resultSetMetaData, "DeviceId,DeviceType");
-      while (resultSet.next()) {
-        String deviceId = resultSet.getString(1);
-        String deviceType = resultSet.getString(2);
-        Assert.assertEquals(expectedDeviceIdList.remove(0), deviceId);
-        Assert.assertEquals(expectedDeviceTypeList.remove(0), deviceType);
-      }
-    }
-  }
-}
